requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0
torch>=2.1.0
transformers>=4.40.0

# ── Vector Database ───────────────────────────────────────────────────────────
endee>=0.1.0

# ── Web scraping ──────────────────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── LLM providers (install the one you need) ─────────────────────────────────
openai>=1.30.0               # For LLM_PROVIDER=openai
google-generativeai>=0.7.0   # For LLM_PROVIDER=google

# ── Evaluation & numerics ─────────────────────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.24.0

# ── UI ────────────────────────────────────────────────────────────────────────
gradio>=4.36.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
huggingface-hub>=0.23.0
